Implement SystemHost methods in header file: Init, Update, Shutdown, and SortSystems; remove SystemHost.cpp
